code refactored, useless components deleted
The entire Menu logic was changed, it is much more effective now
