+++ src/App.js
@@ -1,4 +1,6 @@
-import {Routes, Route, useLocation, Navigate, useNavigate, useParams} from 'react-router-dom';
+import {Routes, Route, useLocation} from 'react-router-dom';
+import { useState, useEffect } from 'react';
+import { useSelector } from 'react-redux';
 import ConsumerCal from './pages/ConsumerCal';
 import NotFound from './pages/NotFound';
 import EventSuccess from './components/EventSuccess';
@@ -6,23 +8,21 @@ import ConsumerEvent from './pages/ConsumerEvent';
 import EventComplete from './pages/EventComplete';
 import ProviderEvents from './pages/ProviderEvents';
 import Account from './pages/Account';
-import Dashboard from './pages/Dashboard';
-import { useSelector } from 'react-redux';
-import LoginPage from './pages/LoginPage';
-import { useState, useEffect } from 'react';
 import SignUpPage from './pages/SignUpPage';
 import Wizard from './components/Wizard';
-import { current } from 'immer';
+import LeftNavbar from './components/LeftNavbar'
+import TopNavbar from './components/TopNavbar';
+import Login from './components/Login';
 
 export default function App() {
   
-  const location=useLocation()
+  const location=(useLocation()).pathname
 
   const [state, setState] = useState({
     providers:[],
     provider:null,
     link:null,
-    path:location.pathname.substring(1)
+    path:location.substring(1)
   })
   
 
@@ -59,13 +59,21 @@ export default function App() {
   const loggedUser = localStorage.getItem('user')
   const currentDate = useSelector(state=>state.chosenDate)
   
+  function menu() {
+    if (location!=='/wizard') return true
+    else return false
+  }
+
+
   return (
     <div>
+      {menu()? <LeftNavbar/> : null}
+      {menu()? <TopNavbar loggedUser={loggedUser}/> : null}
+      <div className={menu()? 'providerContent' : null}>
       <Routes>
-      <Route path="/dashboard" element={<Dashboard/>}/>
       <Route path="/events" element={<ProviderEvents/>}/>
       <Route path="/account" element={<Account/>}/>
-      <Route path="/login" element={<LoginPage/>}/>
+      <Route path="/login" element={<Login/>}/>
       <Route path="/signup" element={<SignUpPage/>}/>
       <Route path="/success" element={<EventSuccess/>}/>
       <Route path="/wizard" element={<Wizard/>}/>
@@ -74,10 +82,10 @@ export default function App() {
       <Route exact path={`/${state.link}/:type/`} element={<ConsumerCal provider={state.provider} />}/>
       <Route exact path={`/${state.link}/:type/${currentDate}`} element={<EventComplete provider={state.provider} />}/>
 
-      <Route path="/" exact element={<Dashboard loggedUser={loggedUser}/>}/>
+      <Route path="/" exact element={<NotFound loggedUser={loggedUser}/>}/>
       <Route path="*" exact element={<NotFound/>}/>  
       </Routes>
-      
+      </div>
     </div>
   )
 }
