+++ src/components/ProviderNavbar.js
@@ -1,17 +1,24 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { NavLink, useLocation } from 'react-router-dom'
 
 export default function ProviderNavbar(props) {
   
   const loggedUser = props.loggedUser
+  const location = (useLocation()).pathname
+
+  useEffect (()=> {
+    console.log(location)
+  })
+
 
   return (
     <div className="provNav">
       <div className="providerNavLogo">Calendar.io</div>
-      {loggedUser && <NavLink className="providerNavButton" to="/dashboard">Dashboard</NavLink>}
-      {loggedUser && <NavLink className="providerNavButton" to='/events'>Events</NavLink>}
-      {loggedUser && <NavLink className="providerNavButton" to='/account'>Account</NavLink>}            
-                                                                        
+      {loggedUser && <div className="column">
+        <NavLink className={location==='/dashboard'? 'menu_button_active':'providerNavButton'} to="/dashboard">Dashboard</NavLink>
+        <NavLink className={location==='/events'? 'menu_button_active':'providerNavButton'} to='/events'>Events</NavLink>
+        <NavLink className={location==='/account'? 'menu_button_active':'providerNavButton'} to='/account'>Account</NavLink>                                                           
+      </div>}
     </div>
   )
 }
