+++ src/index.css
@@ -479,16 +479,22 @@ code {
 
 .providerContent {
   margin-left: 250px;
-  margin-top: 60px;
+  margin-top: 52px;
 }
 
-.providerNavButton {
+.providerNavButton, .menu_button_active {
   font-style: none;
   text-decoration: none;
+  width:200px;
   padding:5px;
   margin-top:30px;
   color:black;
 }
+
+.menu_button_active {
+  background-color: rgb(187, 187, 187);
+}
+
 .providerNavButton:hover {
   background-color: rgb(187, 187, 187);
 }
@@ -673,4 +679,12 @@ cursor: pointer;
 
  .wizard_button:hover {
    cursor:pointer;
+ }
+
+ .flexible_text {
+
+ }
+
+ .flexible_text_active {
+   
  }
