+++ src/pages/SignUpPage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ProviderNavbar from '../components/ProviderNavbar'
-import ProviderTopNavbar from '../components/ProviderTopNavbar'
+import ProviderTopNavbar from '../components/TopNavbar'
 import SignUp from '../components/SignUp'
 
 export default function Account() {
